Export `start_kernel()` for tests
